TraceRA: outsource IntervalBuilderUtil.
